Update file names and class names
